--- ota-webapp/src/main/java/org/tudelft/blockchain/booking/otawebapp/util/Util.java
@@ -1,53 +1,0 @@
-package org.tudelft.blockchain.booking.otawebapp.util;
-
-import org.springframework.stereotype.Component;
-import org.tudelft.blockchain.booking.otawebapp.model.hyperledger.CAEnrollment;
-
-import javax.xml.bind.DatatypeConverter;
-import java.io.*;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.security.KeyFactory;
-import java.security.NoSuchAlgorithmException;
-import java.security.PrivateKey;
-import java.security.spec.InvalidKeySpecException;
-import java.security.spec.PKCS8EncodedKeySpec;
-
-@Component
-public class Util {
-
-    public static CAEnrollment getEnrollment(String keyFolderPath, String keyFileName, String certFolderPath, String certFileName)
-            throws IOException, NoSuchAlgorithmException, InvalidKeySpecException {
-        PrivateKey key = null;
-        String certificate = null;
-        InputStream isKey = null;
-        BufferedReader brKey = null;
-
-        try {
-
-            isKey = new FileInputStream(keyFolderPath + File.separator + keyFileName);
-            brKey = new BufferedReader(new InputStreamReader(isKey));
-            StringBuilder keyBuilder = new StringBuilder();
-
-            for (String line = brKey.readLine(); line != null; line = brKey.readLine()) {
-                if (line.indexOf("PRIVATE") == -1) {
-                    keyBuilder.append(line);
-                }
-            }
-
-            certificate = new String(Files.readAllBytes(Paths.get(certFolderPath, certFileName)));
-
-            byte[] encoded = DatatypeConverter.parseBase64Binary(keyBuilder.toString());
-            PKCS8EncodedKeySpec keySpec = new PKCS8EncodedKeySpec(encoded);
-            KeyFactory kf = KeyFactory.getInstance("EC");
-            key = kf.generatePrivate(keySpec);
-        } finally {
-            isKey.close();
-            brKey.close();
-        }
-
-        CAEnrollment enrollment = new CAEnrollment(key, certificate);
-        return enrollment;
-    }
-    
-}
