+++ ota-webapp/src/test/java/org/tudelft/blockchain/booking/otawebapp/service/CredentialServiceTest.java
@@ -2,21 +2,13 @@ package org.tudelft.blockchain.booking.otawebapp.service;
 
 import org.hyperledger.fabric.sdk.Enrollment;
 import org.hyperledger.fabric.sdk.User;
-import org.hyperledger.fabric.sdk.identity.X509Enrollment;
-import org.hyperledger.fabric.sdk.user.IdemixUser;
 import org.hyperledger.fabric_ca.sdk.exception.RegistrationException;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
-import org.tudelft.blockchain.booking.otawebapp.util.Util;
-
-import java.io.File;
-import java.security.PrivateKey;
 
 import static org.junit.Assert.*;
